DDCNLS-110: Refactored/renamed documentation tests.
